[core] library: run_parallel: move the Parallel code to a dedicate file
